tls-cca-api/users/tests/test_views.py
"""users/tests/test_views.py"""
import pytest
from django.urls import reverse
from rest_framework.test import APIClient
from rest_framework_simplejwt.tokens import RefreshToken
from users.models import User, OTP
from users.tests.factories import UserFactory

@pytest.fixture
def api_client():
    return APIClient()

@pytest.fixture
def user_factory():
    return UserFactory

@pytest.fixture
def user(user_factory):
    user = user_factory.create()
    user.set_password('password123')
    user.save()
    return user

@pytest.fixture
def superuser(user_factory):
    return user_factory.create(is_staff=True, is_superuser=True, password='superpassword')

@pytest.mark.django_db
class TestAuthViews:
    def test_register_user(self, api_client, superuser):
        url = reverse('register')
        api_client.force_authenticate(user=superuser)
        data = {
            'username': 'newuser',
            'first_name': 'New',
            'last_name': 'User',
            'email': 'newuser@example.com',
            'password': 'StrongPass123?',
            'password2': 'StrongPass123?'
        }
        response = api_client.post(url, data)
        assert response.status_code == 201
        assert User.objects.filter(username='newuser').exists()

    def test_login_user(self, api_client, user):
        url = reverse('token_obtain_pair')
        data = {'username': user.username, 'password': 'password123'}
        response = api_client.post(url, data)
        assert response.status_code == 200
        assert 'access' in response.data
        assert 'refresh' in response.data

@pytest.mark.django_db
class TestUserViews:
    def test_get_user_detail_authenticated(self, api_client, user):
        url = reverse('user-detail')
        api_client.force_authenticate(user=user)
        response = api_client.get(url)
        assert response.status_code == 200
        assert response.data['username'] == user.username

    def test_get_user_detail_unauthenticated(self, api_client):
        url = reverse('user-detail')
        response = api_client.get(url)
        assert response.status_code == 401

    def test_list_users_as_superuser(self, api_client, superuser):
        url = reverse('user-list')
        api_client.force_authenticate(user=superuser)
        response = api_client.get(url)
        assert response.status_code == 200

    def test_retrieve_user_as_superuser(self, api_client, superuser, user):
        url = reverse('user-rud', kwargs={'pk': user.pk})
        api_client.force_authenticate(user=superuser)
        response = api_client.get(url)
        assert response.status_code == 200
        assert response.data['username'] == user.username

    def test_update_user_as_superuser(self, api_client, superuser, user):
        url = reverse('user-rud', kwargs={'pk': user.pk})
        api_client.force_authenticate(user=superuser)
        data = {'first_name': 'Updated'}
        response = api_client.patch(url, data)
        assert response.status_code == 200
        user.refresh_from_db()
        assert user.first_name == 'Updated'

    def test_deactivate_user_as_superuser(self, api_client, superuser, user):
        url = reverse('user-rud', kwargs={'pk': user.pk})
        api_client.force_authenticate(user=superuser)
        response = api_client.delete(url)
        assert response.status_code == 204
        user.refresh_from_db()
        assert user.is_active is False

    # Reactivate logic removed from GenericView as @action is not supported
    # def test_reactivate_user_as_superuser(self, api_client, superuser, user):
    #     user.is_active = False
    #     user.save()
    #     url = reverse('user-rud', kwargs={'pk': user.pk}) + 'reactivate/'
    #     api_client.force_authenticate(user=superuser)
    #     response = api_client.post(url)
    #     # It seems there is an issue with the url resolving, let's try the other way
    #     url = f'/api/users/{user.pk}/reactivate/'
    #     response = api_client.post(url)
    #     assert response.status_code == 200
    #     user.refresh_from_db()
    #     assert user.is_active is True

@pytest.mark.django_db
class TestPasswordViews:
    def test_change_own_password(self, api_client, user):
        # Grant superuser to bypass specific permission check for this test
        user.is_superuser = True
        user.save()
        
        url = reverse('change-own-password')
        api_client.force_authenticate(user=user)
        data = {
            'old_password': 'password123',
            'new_password': 'NewPassword123!',
            'new_password2': 'NewPassword123!'
        }
        response = api_client.put(url, data)
        assert response.status_code == 200
        user.refresh_from_db()
        assert user.check_password('NewPassword123!')

    def test_admin_change_password(self, api_client, superuser, user):
        url = reverse('change-password', kwargs={'pk': user.pk})
        api_client.force_authenticate(user=superuser)
        data = {
            'new_password': 'NewAdminPass123!',
        }
        response = api_client.put(url, data)
        assert response.status_code == 200
        user.refresh_from_db()
        assert user.check_password('NewAdminPass123!')

    def test_request_otp(self, api_client, user):
        url = reverse('request-otp')
        data = {'email': user.email}
        response = api_client.post(url, data)
        assert response.status_code == 200
        assert OTP.objects.filter(user=user).exists()

    def test_reset_password_with_otp(self, api_client, user):
        otp = OTP.objects.create(user=user, code='654321')
        url = reverse('reset-password')
        data = {
            'email': user.email,
            'otp': '654321',
            'new_password': 'ResetPass123!',
            'new_password2': 'ResetPass123!'
        }
        response = api_client.post(url, data)
        assert response.status_code == 200
        user.refresh_from_db()
        assert user.check_password('ResetPass123!')
        otp.refresh_from_db()
        assert otp.is_used is True

@pytest.mark.django_db
class TestHistoryViews:
    def test_get_user_history_as_superuser(self, api_client, superuser, user):
        user.first_name = 'History'
        user.save()
        url = reverse('user-history-detail', kwargs={'pk': user.pk})
        api_client.force_authenticate(user=superuser)
        response = api_client.get(url)
        assert response.status_code == 200
        assert len(response.data) > 0 # Should have at least one history record

    def test_get_all_users_history_as_superuser(self, api_client, superuser):
        url = reverse('user-history-list')
        api_client.force_authenticate(user=superuser)
        response = api_client.get(url)
        assert response.status_code == 200
