tls-cca-api/pyproject.toml
[tool.poetry]
name = "tls-cca-api"
version = "0.1.0"
description = ""
authors = ["KJD <125159447+Thrylos77@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.4"
python-dotenv = "1.0.0"
djangorestframework = "^3.16.0"
djangorestframework-simplejwt = "^5.5.0"
drf-spectacular = "^0.28.0"
drf-spectacular-sidecar = "^2025.7.1"
django-simple-history = "^3.10.1"
pytest-cov = "^7.0.0"
django-filter = "^25.2"
psycopg = "^3.3.2"
pytest-django = "^4.11.1"
faker = "^40.1.0"
factory-boy = "^3.3.3"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py"]
addopts = "-ra"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
